moving the getdropdowndata method to its proper place
