+++ postgresConnector.py
@@ -4,10 +4,6 @@ This class extends the base database utility class and is used to establish the 
 
 import sys
 import sqlalchemy as sql
-import pandas as pd
-import os
-import imp
-import configparser as cp
 from baseDbUtil import BaseDbUtil
 
 sys.path.append('.')
